also generate old/armored format for rpm distros
So far the keyring was the concatenation of the binary public keys. This
works fine for newer distros, but makes problems on older ones like
RHEL7 (which should not be used anyways anymore). Still, in the RHEL
universe we are the unicorn here (even on RHEL9/10), so use the armored
version for these.

gpg is one of the most horrible pieces of software, so don't convert "on
the fly", just commit the asc versions to this repo as well.